src/web/tictactoe/backend/requirements.txt
cobs
websocket-client==0.44.0
pyserial